push logger

+++ eval.py
@@ -8,7 +8,7 @@ from tqdm import tqdm
 os.environ["TOKENIZERS_PARALLELISM"] = "false"
 
 
-def get_model_list(path: str = "model/hf_model.csv", search_key_word: list[str]= None):
+def get_model_list(path: str = "model/hf_model1.csv", search_key_word: list[str]= None):
     """ 
     Inputs:
         path: csv_file contain config 
@@ -84,8 +84,10 @@ def evaluate(model_names: list[str]= None, data_names: list[str]= None,
         for data_path in tqdm(data_list):
             for process_type in process_types:
                 log_path = "logger" + "/" + extract_name(model_path, data_path) + "_" + process_type
+                if os.path.exists((log_path)):
+                    continue
                 os.makedirs(log_path, exist_ok=True)
-                res, score = model.predict_dataframe(data_path, process_type= process_type, size= size,prompt_prefix= prompt_prefix, prompt_suffix= prompt_suffix, intermediate_data_save_path= intermediate_data_save_path)
+                res, score = model.predict_dataframe(data_path, process_type= process_type,prompt_prefix= prompt_prefix, prompt_suffix= prompt_suffix, intermediate_data_save_path= intermediate_data_save_path)
                 res.to_csv(f"{log_path}/result.csv")
                 with open(f"{log_path}/score.txt", "w") as f:
                     f.write(str(score))
@@ -96,13 +98,13 @@ def evaluate(model_names: list[str]= None, data_names: list[str]= None,
 def main():
     # prompt_prefix = """Dựa vào trí nhớ của bạn về các bộ dữ liệu, hãy điền vào đoạn <MASKED> trong câu sau để hoàn thành 1 câu hỏi trắc nghiệm. """
     # prompt_suffix = ""
-    os.environ["CUDA_VISIBLE_DEVICES"]="0,1"
+    os.environ["CUDA_VISIBLE_DEVICES"]="4,5"
     DEVICE = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-    BATCH_SIZE = 16
+    print(DEVICE)
+    BATCH_SIZE = 32
     evaluate(model_names= None, data_names = ["eng_final_domain", "domain_addition_plus"],  process_types= ["mask_wrong_answer", "mask_half_question"], 
-             prompt_prefix= None, prompt_suffix= None, size= None,
+             prompt_prefix= None, prompt_suffix= None,
              intermediate_data_save_path= None, DEVICE= DEVICE, BATCH_SIZE= BATCH_SIZE)
             
-    
 if __name__ == "__main__":
     main()
